updated theme-link shortcode
Updated the theme-link shortcode so that the information it displays is based on the style.css file, or at least what WordPress is seeing from it.

This current version automatically lists the child theme atributes, and a link to the parent theme.

Future versions of this shortcode it would be nice to have the option to turn on or off parent theme information, or even be able to just list the parent, or just the child theme, or both.