Changed check for child themes
When checking for child themes, or more when adding information about parent themes, I was simply checking the "truth" of the variable for the parent theme information.

Currently we are checking for if "isset" for the parent information.

At first I was thinking just put it in "automatically" while getting the information about the parent theme.  But maybe it is better to handle it the way we are.  Except, if we are not going to display parent theme information we might as well not even get the parent theme information.

That should be fairly simple to setup.  Though not sure.