streamlined logic for parent theme listing.
The parent theme will be listed by default, and not if the 'parent' attribute is set to anything other than 'true'.

We need to get it so that we can set the 'parent' attribute to something which evaluates as 'true' even if it is True, TRuE or whatever case.  A case insensitive comparison.  Need to look to see what will work there.