Changed wording of "theme link" shortcode
There is a bit of an issue which I ran into with the way the theme-link is worded.

Themes are not supposed to have the word theme in them.  So for some theme which has that in it, then this change will be partially broken.

If the theme does not have that in it (like Datse Multimedia Base) then it didn't read correctly.

So I added "theme" as a word outside the theme name and link.

Further, I added a space based on the existence of a child theme.  This will mean that if the theme is a child theme (like the currently named "Datse Multimedia Theme") the parent theme information should be displaying more correctly now.

That should be working now.  Let's test it.