Updated style.css for version 0.1.6 release
A couple of changes.  The theme version has been updated, and the theme author is listing Datse Multimedia Productions as one of the listed authors.

There is an issue with the way that this ends up getting handled.  If I want to list a number of authors, I have to list them manually, and I think that I still also have to list an "Author URI".  I will look further into it.