Added <br /> after image
In prior testing with narrower images, the image will have the blog
title start to dispay beside the image.

We probably want to get this to display differently in the future,
though this is the easiest that I know of to handle the image right now
(I think it would be better for it to be a background image to the blog
name/description portion, but that may be difficult to implement without
tearing code appart, and putting it back together again.